refactor(interpreter): extract brace expansion into submodule
Move expand_braces/expand_braces_capped/try_expand_range/split_brace_items
out of the 15k-line interpreter/mod.rs into interpreter/brace_expansion.rs.
Behavior-preserving: child module accesses the private Interpreter struct;
externally-called methods marked pub(super). First slice of the interpreter
decomposition.
